refactor: expenseNotebook package
