+++ backend/src/main/java/org/tarjetamish/expenseNotebook/mapper/IExpenseNotebookConverter.java
@@ -1,0 +1,9 @@
+package org.tarjetamish.expenseNotebook.mapper;
+
+import org.tarjetamish.expenseNotebook.dto.ExpenseNotebookDTO;
+import org.tarjetamish.expenseNotebook.model.ExpenseNotebook;
+
+public interface IExpenseNotebookConverter {
+    ExpenseNotebookDTO toExpenseNotebookDTO(ExpenseNotebook expenseNotebook);
+    ExpenseNotebook toExpenseNotebook(ExpenseNotebookDTO expenseNotebookDTO);
+}
