+++ backend/src/main/java/org/tarjetamish/expenseNotebook/repository/impl/JdbcExpenseNoteRepository.java
@@ -5,7 +5,7 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Repository;
 import org.tarjetamish.expenseNotebook.model.ExpenseNotebook;
 import org.tarjetamish.expenseNotebook.repository.ExpenseNotebookRepository;
-import org.tarjetamish.expenseNotebook.mapper.ExpenseNotebookRowMapper;
+import org.tarjetamish.expenseNotebook.mapper.impl.ExpenseNotebookRowMapper;
 
 import java.util.List;
 import java.util.Optional;
@@ -15,21 +15,22 @@ import java.util.Optional;
 public class JdbcExpenseNoteRepository implements ExpenseNotebookRepository {
 
     private final JdbcTemplate jdbc;
+    private final ExpenseNotebookRowMapper expenseNotebookRowMapper;
     @Override
     public List<ExpenseNotebook> findAll() {
         String sql = "SELECT * FROM tarjeta_mish.expense_notebook";
-        return jdbc.query(sql, new ExpenseNotebookRowMapper());
+        return jdbc.query(sql, expenseNotebookRowMapper);
     }
     @Override
     public Optional<ExpenseNotebook> findByCategory(String category) {
         String sql = "SELECT * FROM tarjeta_mish.expense_notebook WHERE category = ?";
-        return Optional.ofNullable(jdbc.queryForObject(sql, new ExpenseNotebookRowMapper(), category));
+        return Optional.ofNullable(jdbc.queryForObject(sql, expenseNotebookRowMapper, category));
     }
 
     @Override
     public ExpenseNotebook save(ExpenseNotebook expenseNotebook) {
         String sql = "INSERT INTO tarjeta_mish.expensebook (description, idcategorybook, idmovement, iduser, name) VALUES (?, ?, ?, ?, ?)";
-        return jdbc.queryForObject(sql, new ExpenseNotebookRowMapper(), expenseNotebook.getDescription(), expenseNotebook.getCategoryBook().name(), expenseNotebook.getTransaction().name(), expenseNotebook.getIdUser(), expenseNotebook.getName());
+        return jdbc.queryForObject(sql, expenseNotebookRowMapper, expenseNotebook.getDescription(), expenseNotebook.getCategoryBook().name(), expenseNotebook.getTransaction().name(), expenseNotebook.getIdUser(), expenseNotebook.getName());
     }
 
     @Override
